Move backup functions inside of the module to make them accessible for import in other python scripts after installation.
